# perf: use mkdirPromise instead of mkdirpPromise