… packages/yarnpkg-pnp/sources/hook.js
… packages/yarnpkg-pnpify/sources/generateSdk.ts