**What's the problem this PR addresses?**

Yarn currently uses `mkdirpPromise` for creating folders, `mkdirpPromise` offers stabalized `utime` and `chmod` but this isn't always necessary meaning it can be skipped to improve performance.

Continuation of https://github.com/yarnpkg/berry/pull/1586

**How did you fix it?**

Replaced the use of `mkdirpPromise(x)` with `mkdirPromise(x, {recursive: true})` which uses the native mkdir from node which is much faster.

**Result**

Testing on https://github.com/microsoft/TypeScript-Website this reduces the time to fetch from a hot global cache to the local cache by ~2 seconds, from ~7.5 to ~5.5, as well the link step by ~13s, from ~70s to ~57s.

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.